Move definition test to validators tests
